# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Client Portal API for freelance design business
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/files:
    get:
      operationId: listProjectFiles
      tags: [files]
      summary: List files for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

    post:
      operationId: uploadProjectFile
      tags: [files]
      summary: Upload a file to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadFileBody"
      responses:
        "201":
          description: Uploaded file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Comments
  /projects/{projectId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Clients (admin only)
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client with their project history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Dashboard
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, client]
        company:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        company:
          type: string
      required:
        - name
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, review, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        clientId:
          type: integer
        clientName:
          type: string
        budget:
          type: ["number", "null"]
        deadline:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - priority
        - clientId
        - clientName
        - createdAt
        - updatedAt
...
    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        url:
          type: string
        size:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        uploadedBy:
          type: integer
        uploaderName:
          type: string
        uploaderRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - url
        - uploadedBy
        - uploaderName
        - uploaderRole
        - createdAt
    UploadFileBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        size:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
      required:
        - name
        - url
    Comment:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - content
        - authorId
        - authorName
        - authorRole
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: ["string", "null"]
        projectCount:
          type: integer
        activeProjects:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - projectCount
        - activeProjects
        - createdAt
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/Project"
          required:
            - projects
    ClientListResponse:
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - clients
        - total
        - page
        - limit
...