# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Client Portal API for freelance design business
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: Date;
}